requirements.txt
chess
Pillow
numpy
matplotlib
python-dotenv
requests
openai